Implmenteret FloatID(Fra Double). Listen virker, og vi har tilføjet kvm, osv). Nu virker databasen

+++ Green home/Servicess/EjendommeRepository_DB.cs
@@ -100,10 +100,10 @@ namespace Green_home.Services
         {
             Ejendomme e = new Ejendomme();
             e.Id = reader.GetInt32(0);
-            e.Pris = reader.GetFloat(1); // Changed to GetFloat for price
+            e.Pris = (float)reader.GetDouble(1);
             e.Kvm = reader.GetInt32(2);
             e.Energimærke = reader.GetString(3);
-            e.By_id = reader.GetInt32(4); // Corrected property name
+            e.By_id = reader.GetInt32(4); 
             return e;
         }
     }
